Remove dependency on common-sequence
There is little need to use common-sequence as a dependency since the implementation is very small. With iteration over for loops the implementation is even smaller.

This change adds a `commonSequence` function that does the same thing the dependency did and removes the need to bring in even more 3rd party code into this repo.